fix/2677 web route executes raw sql (#2759)
* refactor(stats): move raw SQL out of web route into StatsManager in conductor-core (#2677)

Introduce StatsManager in conductor-core/src/stats.rs with a
theme_unlock_stats() method containing all six queries verbatim.
Reduce the web handler to a single manager call, eliminating the
layer violation. Update openapi.rs to import ThemeUnlockStats
directly from conductor_core::stats.

* fix(lint): correct formatting in stats.rs and openapi.rs imports

---------
