+++ conductor-core/src/lib.rs
@@ -52,6 +52,7 @@ pub mod repo;
 pub(crate) mod retry;
 pub mod runtime;
 pub mod schema_config;
+pub mod stats;
 pub mod text_util;
 pub mod ticket_source;
 pub mod tickets;
